feat(keymap): Update keymap

+++ keyboards/crkbd/keymaps/gen740/dvorak_keydata.h
@@ -17,7 +17,7 @@ struct dvorak_node {
   const uint8_t bounds;  //  First 4 bits: modifier, Last 4 bits: keycode
                          //  bounds & 0xF0: Enter event end of the string
                          //  bounds & 0x0F: Exit event end of the string
-  const uint8_t keys[];
+  const uint16_t keys[];
 };
 
 const dvorak_node_t dvorak_node_root;
